Fix additional compiler warnings in perf.h and demo files
Fixed warnings in:
- perf.h: Initialization and size_t→double conversion warnings
  * Changed `static perfState lps = {{{0}}};` to `{0}`
  * Added explicit (double) casts for size_t divisions (lines 145, 157, 159)

- varintChained.h: Sign conversion in macros
  * Changed literal `1` to `1U` in ternary operators

- varintExternalBigEndian.h: Integer truncation in macros
  * Added (uint8_t) casts for all uint64_t→uint8_t assignments

- varintCompare.c: Multiple sign/truncation conversions
  * Fixed GIVE_XZ, SETUP, ACCOUNT_LOOP, ACCOUNT_FINAL macros
  * Added explicit casts for randInt(), memset sizes, min assignment
  * Fixed varintChainedSimpleEncode32 call with (uint32_t) cast

Compiler status after fixes:
- GCC: 24/25 files clean (96%) - only varintDimension.c fails (f16c)
- Clang: 22/25 files clean (88%) - varintDimension.c, varintPackedTest.c, varintCompare.c

Improvements:
- varintCompare.c: Reduced from 235 warnings to 0 on GCC, ~65 on Clang
- varintPackedTest.c: Now passes on GCC (Clang needs varintPacked.h macro fixes)
- All tests still pass with all sanitizers

+++ src/varintExternalBigEndian.h
@@ -31,16 +31,16 @@ uint64_t varintExternalBigEndianGet(const void *p, varintWidth encoding);
         uint8_t *__restrict _vimp_dst = (uint8_t *)(dst);                      \
         switch (encoding) {                                                    \
         case VARINT_WIDTH_8B:                                                  \
-            _vimp_dst[0] = (val);                                              \
+            _vimp_dst[0] = (uint8_t)(val);                                     \
             break;                                                             \
         case VARINT_WIDTH_24B:                                                 \
-            _vimp_dst[2] = (val)&0xff;                                         \
-            _vimp_dst[1] = ((val) >> 8) & 0xff;                                \
-            _vimp_dst[0] = ((val) >> 16) & 0xff;                               \
+            _vimp_dst[2] = (uint8_t)((val)&0xff);                              \
+            _vimp_dst[1] = (uint8_t)(((val) >> 8) & 0xff);                     \
+            _vimp_dst[0] = (uint8_t)(((val) >> 16) & 0xff);                    \
             break;                                                             \
         case VARINT_WIDTH_16B:                                                 \
-            _vimp_dst[1] = (val)&0xff;                                         \
-            _vimp_dst[0] = ((val) >> 8) & 0xff;                                \
+            _vimp_dst[1] = (uint8_t)((val)&0xff);                              \
+            _vimp_dst[0] = (uint8_t)(((val) >> 8) & 0xff);                     \
             break;                                                             \
         default:                                                               \
             varintExternalBigEndianPutFixedWidth((dst), (val), (encoding));    \
